Pass environment needed for systemd-ask-pass
systemd-ask-pass allows the pin to be provided by the
systemd-credentials mechanism, but that requires the environment
variable $CREDENTIALS_DIRECTORY to be set. Openvpn when launched
by systemd does have this environment variable available, but it
doesn't pass it through to systemd-ask-pass meaning that
systemd-ask-pass is unable to load the pin from
systemd-credentials.

This change passes the $CREDENTIALS_DIRECTORY environment variable
to the execed systemd-ask-pass invocation so that
systemd-credentials can be used for non-interactive pin entry.

Signed-off-by: William <william@blackhats.net.au>